requirements.txt
streamlit>=1.30
pandas>=2.0
altair>=5.0
pyyaml>=6.0
pytz
